refactor(database): map keyVersion to int type
- Map `keyVersion` column in `EncryptedSecretTable` to Kotlin `Int` type
- Introduce `IntColumnAdapter` for `keyVersion` during database creation
- Refactor `LocalDatabaseProvider` to centralize database instance creation
- Update `EncryptedSecretLocalDaoImpl` to use `Int` directly for `keyVersion`
- Add `sqldelight-primitive-adapters` dependency

+++ app/src/commonMain/kotlin/eu/torvian/chatbot/app/database/dao/EncryptedSecretLocalDaoImpl.kt
@@ -35,7 +35,7 @@ class EncryptedSecretLocalDaoImpl(
         queries.insert(
             encryptedSecret = encryptedSecret,
             encryptedDEK = encryptedDEK,
-            keyVersion = keyVersion.toLong(),
+            keyVersion = keyVersion,
             createdAt = createdAt,
             updatedAt = updatedAt
         ).let { rowsUpdated ->
@@ -68,7 +68,7 @@ class EncryptedSecretLocalDaoImpl(
             queries.update(
                 encryptedSecret = encryptedSecret,
                 encryptedDEK = encryptedDEK,
-                keyVersion = keyVersion.toLong(),
+                keyVersion = keyVersion,
                 updatedAt = updatedAt,
                 id = id
             ).let { rowsUpdated ->
